--- src/frontend/components/configs/ConfigsPage.jsx
@@ -1,18 +1,0 @@
-import ConfigsList from "./ConfigsList";
-
-export default function ConfigsPage() {
-  return (
-    <div className="py-6">
-      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-        <h1 className="text-2xl font-semibold text-gray-900">
-          All Configurations
-        </h1>
-      </div>
-      <div className="max-w-7xl mx-auto px-4 sm:px-6 md:px-8">
-        <div className="py-4">
-          <ConfigsList />
-        </div>
-      </div>
-    </div>
-  );
-}
