--- src/frontend/components/configs/ConfigurationsSlice.js
@@ -1,7 +1,0 @@
-// import { userAPI } from '../../../../../stats/src/userAPI';
-// import {
-//   createSlice,
-//   createAsyncThunk,
-//   createEntityAdapter,
-// } from '@reduxjs/toolkit';
-
